requirements.txt
Flask
Flask-SQLAlchemy
pandas
psycopg2-binary
python-dotenv